cloud_functions/gemini_analysis/requirements.txt
google-api-python-client==2.92.0
google-auth==2.23.2
google-cloud-storage==2.11.0
Flask==2.3.3
functions-framework==3.*
google-cloud-aiplatform==1.60.0
google-cloud-logging==3.11.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-secret-manager==2.16.2